use the connect method to generate container components
The connect method from the react-redux package will offers a way to
generate a container component by receiving two functions as
parameters. The first one is a map between the state and the props and
the second is a map between the dispatch and props. And finally, we
call the result of the connect function passing the presentation
component as parameter.